src/requirements.txt
# CLI Framework
typer[all]==0.9.0
rich==13.3.5
click==8.1.7

# Environment Management
python-dotenv==1.0.0

# API Clients
notion-client==2.0.0
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-generativeai==0.7.2
whatsapp-api-client-python

# Project dependencies will be listed here.
# e.g., requests==2.28.1 